Reorganize mockk-android tests
The io.mockk.gh package tests in the mockk-android module are moved
and renamed to better describe which features are being tested.
