+++ mockk/android/src/androidTest/java/io/mockk/MethodDescriptionTest.kt
@@ -1,12 +1,14 @@
-package io.mockk.gh
+package io.mockk
 
-import io.mockk.every
-import io.mockk.mockk
 import java.util.concurrent.Callable
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class Issue170Test {
+class MethodDescriptionTest {
+
+    /**
+     * Verifies issue #170.
+     */
     @Test
     fun test() {
         val mock = mockk<Callable<String>>()
