+++ mockk/android/src/androidTest/java/io/mockk/proxy/android/AndroidMockKAgentFactoryTest.kt
@@ -1,4 +1,4 @@
-package io.mockk.gh
+package io.mockk.proxy.android
 
 import android.support.test.rule.ActivityTestRule
 import android.widget.FrameLayout
@@ -7,16 +7,19 @@ import io.mockk.mockk
 import org.junit.Rule
 import org.junit.Test
 
-class Issue563Test {
+class AndroidMockKAgentFactoryTest {
 
     @Rule
     @JvmField
     val rule = ActivityTestRule(TestActivity::class.java)
 
+    /**
+     * This tests that the hidden api logic in [AndroidMockKAgentFactory] works. Otherwise, we would fail when mocking
+     * the FrameLayout.
+     * Verifies issue #563.
+     */
     @Test
     fun test() {
-        // This tests that the hidden api logic in AndroidMockKAgentFactory works. Otherwise,
-        // we would fail when mocking the FrameLayout here
         mockk<FrameLayout>()
     }
 }
